pyproject.toml
[project]
name = "expense-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gspread>=6.2.1",
    "imagehash>=4.3.2",
    "numpy>=2.2.6",
    "openai>=1.107.2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "pyyaml>=6.0.2",
    "streamlit>=1.49.1",
    "tesseract>=0.1.3",
]
